I've abstracted out the common code from the two mongo repos into a base class that can store arbitrary data (as long as there's a json formatter for it) in a collection against a NINO.
