+++ app/uk/gov/hmrc/mobilehelptosave/repository/MongoFeatureFlagsRepo.scala
@@ -16,18 +16,11 @@
 
 package uk.gov.hmrc.mobilehelptosave.repository
 
-import cats.instances.future._
-import cats.syntax.functor._
-import javax.inject.{Inject, Provider}
-import play.api.libs.json.Json._
+import javax.inject.{Inject, Provider, Singleton}
 import play.api.libs.json.{Format, Json, OWrites, Reads}
 import play.modules.reactivemongo.ReactiveMongoComponent
-import reactivemongo.api.indexes.{Index, IndexType}
-import reactivemongo.bson.BSONObjectID
-import uk.gov.hmrc.domain.Nino
-import uk.gov.hmrc.mongo.ReactiveRepository
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.ExecutionContext
 
 case class FeatureFlagsMongoModel(nino: String, savingsTargetsEnabled: Boolean)
 
@@ -39,23 +32,10 @@ object FeatureFlagsMongoModel {
     Format(reads, writes)
 }
 
+@Singleton
 class MongoFeatureFlagsRepo @Inject()(
-  val reactiveMongo: Provider[ReactiveMongoComponent]
+  override val reactiveMongo: Provider[ReactiveMongoComponent]
 )
   (implicit ec: ExecutionContext, mongoFormats: Format[FeatureFlagsMongoModel])
-  extends ReactiveRepository[FeatureFlagsMongoModel, BSONObjectID]("featureFlags", reactiveMongo.get().mongoConnector.db, mongoFormats)
-    with FeatureFlagsRepo {
-
-  override def indexes: Seq[Index] = Seq(
-    Index(Seq("nino" -> IndexType.Text), name = Some("ninoIdx"), unique = true, sparse = true)
-  )
-
-  override def put(flags: FeatureFlagsMongoModel): Future[Unit] =
-    insert(flags).void
-
-  override def get(nino: Nino): Future[Option[FeatureFlagsMongoModel]] =
-    find("nino" -> nino.value).map(_.headOption)
-
-  override def delete(nino: Nino): Future[Unit] =
-    remove("nino" -> nino.value).void
-}
+  extends NinoIndexedMongoRepo[FeatureFlagsMongoModel]("featureFlags", reactiveMongo)
+    with FeatureFlagsRepo
