+++ app/uk/gov/hmrc/mobilehelptosave/repository/NinoIndexedMongoRepo.scala
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2018 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.mobilehelptosave.repository
+
+import cats.instances.future._
+import cats.syntax.functor._
+import javax.inject.Provider
+import play.api.libs.json.Format
+import play.modules.reactivemongo.ReactiveMongoComponent
+import reactivemongo.api.indexes.{Index, IndexType}
+import reactivemongo.bson.BSONObjectID
+import uk.gov.hmrc.domain.Nino
+import uk.gov.hmrc.mongo.ReactiveRepository
+
+import scala.concurrent.{ExecutionContext, Future}
+
+/**
+  * Want to store something in Mongo indexed against a NINO? Just create the model for the
+  * data and subclass this handy helper.
+  */
+class NinoIndexedMongoRepo[T](
+  collectionName:String,
+  val reactiveMongo: Provider[ReactiveMongoComponent]
+)(implicit ec: ExecutionContext, mongoFormats: Format[T])
+  extends ReactiveRepository[T, BSONObjectID](collectionName, reactiveMongo.get().mongoConnector.db, mongoFormats) {
+  override def indexes: Seq[Index] = Seq(
+    Index(Seq("nino" -> IndexType.Text), name = Some("ninoIdx"), unique = true, sparse = true)
+  )
+
+  def put(flags: T): Future[Unit] =
+    insert(flags).void
+
+  def get(nino: Nino): Future[Option[T]] =
+    find("nino" -> nino.value).map(_.headOption)
+
+  def delete(nino: Nino): Future[Unit] =
+    remove("nino" -> nino.value).void
+}
